+++ protected/humhub/modules/tour/controllers/TourController.php
@@ -9,13 +9,9 @@
 namespace humhub\modules\tour\controllers;
 
 use humhub\components\Controller;
-use humhub\modules\space\models\Membership;
-use humhub\modules\space\models\Space;
-use humhub\modules\tour\Module;
+use humhub\modules\tour\models\TourParams;
 use humhub\modules\user\models\User;
 use Yii;
-use yii\web\HttpException;
-use yii\web\Response;
 
 /**
  * TourController
@@ -41,55 +37,21 @@ class TourController extends Controller
      */
     public function actionTourCompleted()
     {
-        /** @var Module $module */
-        $module = Yii::$app->getModule('tour');
+        // get page name from completed tour
+        $page = Yii::$app->request->post('page');
 
-        // get section parameter from completed tour
-        $section = Yii::$app->request->post('section');
-
-        if (!in_array($section, $module->acceptableNames)) {
+        if (!TourParams::isPageAcceptable($page)) {
             return;
         }
 
         // set tour status to seen for current user
-        Yii::$app->getModule('tour')->settings->user()->set($section, 1);
+        Yii::$app->getModule('tour')->settings->user()->set($page, true);
     }
 
     public function actionHidePanel()
     {
         // set tour status to seen for current user
-        Yii::$app->getModule('tour')->settings->user()->set('hideTourPanel', 1);
-    }
-
-    /**
-     *  This is a special case, because we need to find a space to start the tour
-     *
-     * @return Response
-     * @throws HttpException
-     */
-    public function actionStartSpaceTour()
-    {
-        $space = null;
-
-        // Loop over all spaces where the user is member
-        foreach (Membership::getUserSpaces() as $space) {
-            if ($space->isAdmin() && !$space->isArchived()) {
-                // If user is admin on this space, it´s the perfect match
-                break;
-            }
-        }
-
-        if ($space === null) {
-            // If user is not member of any space, try to find a public space
-            // to run tour in
-            $space = Space::findOne(['and', ['!=', 'visibility' => Space::VISIBILITY_NONE], ['status' => Space::STATUS_ENABLED]]);
-        }
-
-        if ($space === null) {
-            throw new HttpException(404, 'Could not find any public space to run tour!');
-        }
-
-        return $this->redirect($space->createUrl('/space/space', ['tour' => true]));
+        Yii::$app->getModule('tour')->settings->user()->set('hideTourPanel', true);
     }
 
     /**
@@ -100,14 +62,15 @@ class TourController extends Controller
         /* @var User $user */
         $user = Yii::$app->user->getIdentity();
 
-        if ($user->id == 1 &&
-            $user->load(Yii::$app->request->post()) &&
-            $user->save(true, ['tagsField']) &&
-            ($profile = $user->profile) &&
-            $profile->load(Yii::$app->request->post()) &&
-            $profile->save(true, ['firstname', 'lastname', 'title', 'birthday', 'birthday_hide_year', 'phone_work', 'mobile'])
+        if (
+            $user->id === 1
+            && $user->load(Yii::$app->request->post())
+            && $user->save(true, ['tagsField'])
+            && ($profile = $user->profile)
+            && $profile->load(Yii::$app->request->post())
+            && $profile->save(true, ['firstname', 'lastname', 'title', 'birthday', 'birthday_hide_year', 'phone_work', 'mobile'])
         ) {
-            Yii::$app->getModule('tour')->settings->contentContainer($user)->set('welcome', 1);
+            Yii::$app->getModule('tour')->settings->contentContainer($user)->set('welcome', true);
             return $this->redirect(['/dashboard/dashboard']);
         }